Implemented preproccesing programs as well as update make file

+++ Lab2/gomezj_vector_ops.hpp
@@ -19,12 +19,12 @@ namespace JpSpace {
     public:
         vectorops(int);
         //sum all elements of array
-        int Sum(std::vector<float> *data);
+        int Sum(std::vector<int> *data);
         
         //subtracts one array from other array by element
-        void Subtract(std::vector<float> *data1, std::vector<float> *data2, std::vector<float> *output);
+        void Subtract(std::vector<int> *data1, std::vector<int> *data2, std::vector<int> *output);
         //divides array elements by value
-        void Division(std::vector<float> *data1, float const_div, std::vector<float> *output);
+        void Division(std::vector<int> *data1, float const_div, std::vector<float> *output);
                 
     };
 }
